% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{place_names}
\alias{place_names}
\title{Place Names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 102511 rows and 8 columns.
}
\usage{
place_names
}
\description{
The meaning of the description codes can be found in \code{place_names_description}.
}
\keyword{datasets}
