% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{railway_stations}
\alias{railway_stations}
\title{UK Railway Stations}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2635 rows and 5 columns.
}
\usage{
railway_stations
}
\description{
UK Railway Stations
}
\keyword{datasets}
