% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{railways}
\alias{railways}
\title{UK Railway Network}
\format{
A data frame with 49484 rows and 7 variables:
\describe{
  \item{\code{fid}}{}
  \item{\code{elr}}{Engineer's Line Reference}
  \item{\code{trid}}{Track ID}
  \item{\code{lat_start}}{Starting latitude}
  \item{\code{lat_end}}{Ending latitude}
  \item{\code{lon_start}}{}
  \item{\code{lon_end}}{}
}
}
\source{
\url{https://www.whatdotheyknow.com/request/latitude_and_longitude_of_all_ra}
}
\usage{
railways
}
\description{
Latitude and longitude of all railway tracks in the United Kingdom.
}
\details{
Data from a \href{https://www.whatdotheyknow.com/request/latitude_and_longitude_of_all_ra}{Freedom of Information request}.
}
\keyword{datasets}
